backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pytesseract==0.3.10
Pillow==9.3.0
pymupdf==1.20.2
whisper==1.0
openai==0.27.8
python-multipart
anthropic>=0.3.0
sentence-transformers==2.2.2
spacy==3.7.2
vaderSentiment==3.3.2
numpy>=1.24.0
aiohttp>=3.8.0
asyncio>=3.4.3
redis==5.0.1
aioredis==2.0.1
librosa==0.10.1
soundfile==0.12.1
torch>=2.0.0
pika==1.3.2
aio-pika==9.3.0
aiosqlite==0.19.0
python-dotenv>=1.0.0
textract
pandas
openpyxl
pytesseract
Pillow
SpeechRecognition
moviepy 